Port masterserver samble to net8
